Stop the SSA differential reinterpreting a double as an f32
`TestSelfHostSSABackendAgreesWithStackMachine`'s `floats` program calls
`f32_bits` on an f64. That is the program #9987 was found through, and the
issue records it as already corrected to `f32_bits(y as f32)` — it was not,
so the new argument check refused it and the aarch64 shard went red.

The cast is the correction the issue names, and it is the program's own bug:
native refuses the uncast form too, and only the self-host never saw it,
because both legs of this differential are the self-host CLI. Nothing else in
the tree calls the float-bits builtins uncast; the self-host's own uses
already spell `x as f32`.

The case compares the SSA backend against the stack machine on one program
rather than against a golden answer, so both legs move together and the
agreement it pins is unchanged.

+++ internal/e2eselfhost/self_host_ssa_backend_test.go
@@ -317,7 +317,7 @@ function convs(n: i32, u: u32, w: i64, v: u64): i64 {
 function bits(x: f64): i64 {
     var b: i64 = f64_bits(x);
     var y: f64 = f64_from_bits(b + 1);
-    var h: i32 = f32_bits(y);
+    var h: i32 = f32_bits(y as f32);
     var zf: f32 = f32_from_bits(h);
     var z: f64 = zf as f64;
     return b + (z as i64) + (h as i64) % 7;
